# api/docs/openapi.yaml
openapi: 3.0.0
info:
    title: 'Zyvo API'
    version: 1.0.0
    description: "\n        A secure Node.js API server built with Express.js that implements \n        security best practices including CORS, CSRF protection, rate limiting, \n        and comprehensive error handling.\n        \n        ## Features\n        - JWT Authentication\n        - Role-based Access Control\n        - Rate Limiting\n        - CSRF Protection\n        - Input Validation\n        - Comprehensive Error Handling\n        \n        ## Security\n        This API implements multiple layers of security:\n        - CORS with configurable origins\n        - CSRF token validation\n        - Rate limiting per IP\n        - Secure HTTP headers via Helmet.js\n        - Input sanitization and validation\n      "
    contact:
        name: 'Zyvo API Support'
        email: support@zyvo.com
        url: 'https://zyvo.com'
    license:
        name: ISC
        url: 'https://opensource.org/licenses/ISC'
servers:
    -
        url: 'http://localhost:3000'
        description: 'Development server'
    -
        url: 'https://api.zyvo.com'
        description: 'Production server'
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'JWT Bearer token authentication'
        csrfToken:
            type: apiKey
            in: header
            name: X-CSRF-Token
            description: 'CSRF token for protection against CSRF attacks'
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        BrandResponse:
            type: object
            properties: {success: {type: boolean}, message: {type: string}, data: {$ref: '#/components/schemas/Brand'}}
...
    responses:
        '400':
            description: 'Bad Request - Invalid input data'
            content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
...
paths:
...
    /api/v1/admin/products:
        post:
            summary: 'Create new product'
            description: 'Create a new product in the system'
            tags: [Admin]
            security: [{bearerAuth: []}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {name: {type: string, description: 'Product name'}, price: {type: number, description: 'Product price'}, category: {type: string, description: 'Product category'}, description: {type: string, description: 'Product description'}, stock: {type: integer, description: 'Initial stock quantity'}}}}}}
            responses: {'201': {description: 'Product created successfully'}, '400': {description: 'Invalid input data'}, '401': {description: 'Admin authentication required'}}
...
    /api/v1/brands:
        post:
            summary: 'Create a new brand'
            tags: [Brands]
            security: [{BearerAuth: []}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, required: [name], properties: {name: {type: string, minLength: 2, maxLength: 100, example: Nike}, description: {type: string, maxLength: 1000, example: 'Leading athletic footwear and apparel brand'}, logo_url: {type: string, format: uri, example: 'https://example.com/nike-logo.png'}, website: {type: string, format: uri, example: 'https://www.nike.com'}, contact_email: {type: string, format: email, example: info@nike.com}}}}}}
            responses: {'201': {description: 'Brand created successfully', content: {application/json: {schema: {$ref: '#/components/schemas/BrandResponse'}}}}, '400': {description: 'Validation error or duplicate brand name'}, '401': {description: 'Unauthorized - Admin access required'}, '500': {description: 'Internal server error'}}
        get:
            summary: 'Get all brands with filtering and pagination'
            tags: [Brands]
            parameters: [{in: query, name: page, schema: {type: integer, minimum: 1, default: 1}, description: 'Page number for pagination'}, {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 100, default: 10}, description: 'Number of items per page'}, {in: query, name: search, schema: {type: string, maxLength: 100}, description: 'Search in brand name and description'}, {in: query, name: sort, schema: {type: string, enum: [name, createdAt, updatedAt], default: name}, description: 'Sort field'}, {in: query, name: order, schema: {type: string, enum: [asc, desc], default: asc}, description: 'Sort order'}, {in: query, name: is_active, schema: {type: boolean}, description: 'Filter by active status (admin only)'}, {in: query, name: include_inactive, schema: {type: boolean, default: false}, description: 'Include inactive brands (admin only)'}]
            responses: {'200': {description: 'List of brands retrieved successfully', content: {application/json: {schema: {$ref: '#/components/schemas/BrandsListResponse'}}}}, '400': {description: 'Invalid query parameters'}, '500': {description: 'Internal server error'}}
...
    /api/v1/categories:
        get:
            summary: 'Get all categories'
            description: 'Retrieve a paginated list of categories with optional filtering'
            tags: [Categories]
            parameters: [{in: query, name: page, schema: {type: integer, minimum: 1, default: 1}, description: 'Page number for pagination'}, {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 100, default: 10}, description: 'Number of items per page'}, {in: query, name: search, schema: {type: string}, description: 'Search term for category name or description'}, {in: query, name: parent_id, schema: {type: string}, description: 'Filter by parent category ID (use ''null'' for root categories)'}, {in: query, name: sort_by, schema: {type: string, enum: [name, createdAt, updatedAt], default: name}, description: 'Field to sort by'}, {in: query, name: sort_order, schema: {type: string, enum: [asc, desc], default: asc}, description: 'Sort order'}, {in: query, name: include_inactive, schema: {type: boolean, default: false}, description: 'Include inactive categories (admin only)'}]
            responses: {'200': {description: 'Categories retrieved successfully', content: {application/json: {schema: {type: object, properties: {success: {type: boolean, example: true}, message: {type: string, example: 'Categories retrieved successfully'}, data: {type: array, items: {$ref: '#/components/schemas/Category'}}, pagination: {$ref: '#/components/schemas/PaginationInfo'}}}}}}, '400': {$ref: '#/components/responses/400'}, '500': {$ref: '#/components/responses/500'}}
        post:
            summary: 'Create a new category'
            description: 'Create a new category (admin only)'
            tags: [Categories]
            security: [{bearerAuth: []}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, required: [name], properties: {name: {type: string, minLength: 1, maxLength: 100, example: Electronics}, description: {type: string, maxLength: 500, example: 'Electronic devices and accessories'}, parent_category: {type: string, description: 'MongoDB ObjectID of parent category', example: 64a1b2c3d4e5f6789abcdef1}, image_url: {type: string, format: uri, example: 'https://example.com/electronics.jpg'}}}}}}
            responses: {'201': {description: 'Category created successfully', content: {application/json: {schema: {type: object, properties: {success: {type: boolean, example: true}, message: {type: string, example: 'Category created successfully'}, data: {$ref: '#/components/schemas/Category'}}}}}}, '400': {$ref: '#/components/responses/400'}, '401': {$ref: '#/components/responses/401'}, '403': {$ref: '#/components/responses/403'}, '500': {$ref: '#/components/responses/500'}}
...
    /api/v1/inventory:
        post:
            summary: 'Create a new inventory record'
            tags: [Inventory]
            security: [{adminAuth: []}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, required: [product_variant_id], properties: {product_variant_id: {type: string, description: 'ID of the base unit product variant'}, stock_quantity: {type: number, minimum: 0, default: 0, description: 'Initial stock quantity'}, min_stock_level: {type: number, minimum: 0, default: 0, description: 'Minimum stock threshold'}, location: {type: string, maxLength: 200, description: 'Physical storage location'}, notes: {type: string, maxLength: 1000, description: 'Internal notes'}}}}}}
            responses: {'201': {description: 'Inventory record created successfully'}, '400': {description: 'Validation error or pack variant provided'}, '401': {description: 'Unauthorized - admin access required'}, '500': {description: 'Internal server error'}}
        get:
            summary: 'Get all inventory records with pagination and filtering'
            tags: [Inventory]
            security: [{adminAuth: []}]
            parameters: [{in: query, name: page, schema: {type: integer, minimum: 1, default: 1}, description: 'Page number for pagination'}, {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 100, default: 10}, description: 'Number of items per page'}, {in: query, name: is_active, schema: {type: boolean}, description: 'Filter by active status'}, {in: query, name: stock_status, schema: {type: string, enum: [out_of_stock, low_stock, in_stock]}, description: 'Filter by stock status'}, {in: query, name: location, schema: {type: string}, description: 'Filter by location (partial match)'}, {in: query, name: product_id, schema: {type: string}, description: 'Filter by product ID'}, {in: query, name: search, schema: {type: string}, description: 'Search across SKU, location, and notes'}, {in: query, name: include_computed_packs, schema: {type: boolean, default: false}, description: 'Include computed stock for pack variants'}, {in: query, name: sort_by, schema: {type: string, enum: [createdAt, updatedAt, stock_quantity, min_stock_level, last_restock_date], default: createdAt}, description: 'Field to sort by'}, {in: query, name: sort_order, schema: {type: string, enum: [asc, desc], default: desc}, description: 'Sort order'}]
            responses: {'200': {description: 'Inventory records retrieved successfully'}, '401': {description: 'Unauthorized - admin access required'}, '500': {description: 'Internal server error'}}
...
    /api/v1/options:
        post:
            summary: 'Create a new option'
            tags: [Options]
            security: [{BearerAuth: []}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, required: [option_type, option_value], properties: {option_type: {type: string, maxLength: 50, example: Color}, option_value: {type: string, maxLength: 100, example: Red}, name: {type: string, maxLength: 100, example: 'Bright Red'}, sort_order: {type: number, minimum: 0, example: 1}}}}}}
            responses: {'201': {description: 'Option created successfully', content: {application/json: {schema: {$ref: '#/components/schemas/OptionResponse'}}}}, '400': {description: 'Validation error or duplicate option'}, '401': {description: 'Unauthorized - Admin access required'}, '500': {description: 'Internal server error'}}
        get:
            summary: 'Get all options with filtering and pagination'
            tags: [Options]
            parameters: [{in: query, name: page, schema: {type: integer, minimum: 1}, description: 'Page number for pagination'}, {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 100}, description: 'Number of items per page'}, {in: query, name: option_type, schema: {type: string, maxLength: 50}, description: 'Filter by option type (case-insensitive)'}, {in: query, name: is_active, schema: {type: boolean}, description: 'Filter by active status (admin only)'}, {in: query, name: search, schema: {type: string, maxLength: 100}, description: 'Search in name, option_value, and option_type'}, {in: query, name: sort, schema: {type: string, enum: [option_type, name, sort_order]}, description: 'Sort field'}, {in: query, name: order, schema: {type: string, enum: [asc, desc]}, description: 'Sort order'}, {in: query, name: include_inactive, schema: {type: boolean}, description: 'Include inactive options (admin only)'}]
            responses: {'200': {description: 'List of options retrieved successfully', content: {application/json: {schema: {$ref: '#/components/schemas/OptionsListResponse'}}}}, '400': {description: 'Invalid query parameters'}, '500': {description: 'Internal server error'}}
...
    /api-docs:
        get:
            summary: 'API Documentation'
            description: "Serves the interactive API documentation generated with Redoc.\nThis endpoint provides access to the complete API documentation \nincluding all endpoints, schemas, and examples.\n"
            tags: [Health]
            responses: {'200': {description: 'API documentation page', content: {text/html: {schema: {type: string, example: '<!DOCTYPE html><html>...</html>'}}}}, '404': {description: 'Documentation not found (may not be generated yet)', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}}